requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
absl-py=0.12.0=pypi_0
antspyx=0.2.7=pypi_0
anyio=2.2.0=py37h89c1867_0
argon2-cffi=20.1.0=py37h4abf009_2
asn1crypto=1.4.0=pyh9f0ad1d_0
astor=0.8.1=pypi_0
async_generator=1.10=py_0
attrs=20.3.0=pyhd3deb0d_0
babel=2.9.0=pyhd3deb0d_0
backcall=0.2.0=pyhd3eb1b0_0
bleach=3.3.0=pyh44b312d_0
brotlipy=0.7.0=py37hb5d75c8_1001
bzip2=1.0.8=h516909a_3
ca-certificates=2020.12.5=ha878542_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cairo=1.16.0=hcf35c78_1003
certifi=2020.12.5=py37h89c1867_1
cffi=1.14.4=py37h11fe52a_0
chardet=4.0.0=py37h89c1867_1
chart-studio=1.1.0=pypi_0
cryptography=2.5=py37hb7f436b_1
cycler=0.10.0=py_2
dbus=1.13.6=he372182_0
decorator=4.4.2=pyhd3eb1b0_0
deepbrain=0.1=pypi_0
defusedxml=0.7.1=pyhd8ed1ab_0
entrypoints=0.3=pyhd8ed1ab_1003
expat=2.2.9=he1b5a44_2
ffmpeg=4.3.1=h3215721_1
fontconfig=2.13.1=h86ecdb6_1001
freetype=2.10.4=h7ca028e_0
gast=0.2.2=pypi_0
gettext=0.19.8.1=hf34092f_1004
glib=2.66.1=h680cd38_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-pasta=0.2.0=pypi_0
graphite2=1.3.13=h58526e2_1001
grpcio=1.36.1=pypi_0
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
h5py=3.2.1=pypi_0
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h3c11f04_101
icu=64.2=he1b5a44_1
idna=2.10=pyh9f0ad1d_0
imageio=2.9.0=pypi_0
importlib-metadata=3.7.3=py37h89c1867_0
ipykernel=5.3.4=py37h5ca1d4c_0
ipython=7.21.0=py37hb070fc8_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=7.6.3=pyhd3deb0d_0
jasper=1.900.1=h07fcdf6_1006
jedi=0.17.0=py37_0
jinja2=2.11.3=pyh44b312d_0
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
json5=0.9.5=pyh9f0ad1d_0
jsonschema=3.2.0=pyhd8ed1ab_3
jupyter=1.0.0=py37h89c1867_6
jupyter-packaging=0.7.12=pyhd8ed1ab_0
jupyter_client=6.1.7=py_0
jupyter_console=6.3.0=pyhd8ed1ab_0
jupyter_core=4.7.1=py37h06a4308_0
jupyter_server=1.4.1=py37h89c1867_0
jupyterlab=3.0.12=pyhd8ed1ab_0
jupyterlab_pygments=0.1.2=pyh9f0ad1d_0
jupyterlab_server=2.3.0=pyhd8ed1ab_0
jupyterlab_widgets=1.0.0=pyhd8ed1ab_1
keras-applications=1.0.8=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.3.1=py37hc928c03_0
lame=3.100=h14c3975_1001
lcms2=2.11=hcbb858e_1
ld_impl_linux-64=2.33.1=h53a641e_7
libblas=3.9.0=8_openblas
libcblas=3.9.0=8_openblas
libclang=9.0.1=default_hde54327_0
libedit=3.1.20191231=h14c3975_1
libffi=3.2.1=hf484d3e_1007
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.5.0=h14aa051_18
libgfortran4=7.5.0=h14aa051_18
libiconv=1.16=h516909a_0
liblapack=3.9.0=8_openblas
liblapacke=3.9.0=8_openblas
libllvm9=9.0.1=he513fc3_1
libopenblas=0.3.12=pthreads_hb3c22a3_1
libopencv=4.5.0=py37_3
libpng=1.6.37=h21135ba_2
libsodium=1.0.18=h7b6447c_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h4f3a223_6
libuuid=2.32.1=h14c3975_1000
libwebp-base=1.1.0=h36c2ea0_3
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
lz4-c=1.9.2=he1b5a44_3
markdown=3.3.4=pypi_0
markupsafe=1.1.1=py37hb5d75c8_2
matplotlib=3.3.4=pypi_0
matplotlib-base=3.3.3=py37h4f6019d_0
mistune=0.8.4=py37h4abf009_1002
nbclassic=0.2.6=pyhd8ed1ab_0
nbclient=0.5.3=pyhd8ed1ab_0
nbconvert=6.0.7=py37h89c1867_3
nbformat=5.1.2=pyhd8ed1ab_1
nbresuse=0.4.0=pyhd8ed1ab_0
ncurses=6.2=he6710b0_1
nest-asyncio=1.4.3=pyhd8ed1ab_0
nettle=3.6=he412f7d_0
networkx=2.5=pypi_0
nibabel=3.2.1=pyhd8ed1ab_0
nilearn=0.7.1=pyhd8ed1ab_0
notebook=6.2.0=py37h89c1867_0
nspr=4.29=he1b5a44_1
nss=3.59=h2c00c37_0
numpy=1.20.1=pypi_0
olefile=0.46=pyh9f0ad1d_1
opencv=4.5.0=py37_3
openh264=2.1.1=h8b12597_0
openssl=1.0.2u=h516909a_0
opt-einsum=3.3.0=pypi_0
packaging=20.9=pyh44b312d_0
pandas=1.1.4=py37h10a2094_0
pandoc=2.12=h7f98852_0
pandocfilters=1.4.2=py_1
parso=0.8.1=pyhd3eb1b0_0
patsy=0.5.1=py_0
pcre=8.44=he1b5a44_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.1.2=pypi_0
pip=21.0.1=py37h06a4308_0
pixman=0.38.0=h516909a_1003
plotly=4.14.3=pypi_0
prometheus_client=0.9.0=pyhd3deb0d_0
prompt-toolkit=3.0.8=py_0
prompt_toolkit=3.0.8=0
protobuf=3.15.6=pypi_0
psutil=5.4.8=py37h14c3975_1000
pthread-stubs=0.4=h36c2ea0_1001
ptyprocess=0.7.0=pyhd3eb1b0_2
py-opencv=4.5.0=py37h888b3d9_3
pycparser=2.20=pyh9f0ad1d_2
pydicom=2.1.2=pyhd3deb0d_0
pygments=2.8.1=pyhd3eb1b0_0
pyopenssl=19.0.0=py37_0
pyparsing=2.4.7=pyh9f0ad1d_0
pyqt=5.12.3=py37h8685d9f_3
pyqt5-sip=4.19.18=pypi_0
pyqtchart=5.12=pypi_0
pyqtwebengine=5.12.1=pypi_0
pyrsistent=0.17.3=py37h4abf009_1
pysocks=1.7.1=py37h89c1867_3
python=3.7.0=h6e4f718_3
python-dateutil=2.8.1=pyhd3eb1b0_0
python_abi=3.7=1_cp37m
pytz=2021.1=pyhd8ed1ab_0
pywavelets=1.1.1=pypi_0
pyyaml=5.4.1=pypi_0
pyzmq=20.0.0=py37h2531618_1
qt=5.12.5=hd8c4c69_1
qtconsole=5.0.3=pyhd8ed1ab_0
qtpy=1.9.0=py_0
readline=7.0=h7b6447c_5
requests=2.25.1=pyhd3deb0d_0
retrying=1.3.3=pypi_0
rope=0.18.0=py_0
scikit-image=0.18.1=pypi_0
scikit-learn=0.23.2=py37hddcf8d6_3
scipy=1.6.1=pypi_0
seaborn=0.11.1=hd8ed1ab_1
seaborn-base=0.11.1=pyhd8ed1ab_1
send2trash=1.5.0=py_0
setuptools=52.0.0=py37h06a4308_0
six=1.15.0=py37h06a4308_0
sklearn=0.0=pypi_0
sniffio=1.2.0=py37h89c1867_1
sqlite=3.33.0=h62c20be_0
statsmodels=0.12.2=pypi_0
tensorboard=1.15.0=pypi_0
tensorflow=1.15.0=pypi_0
tensorflow-estimator=1.15.1=pypi_0
termcolor=1.1.0=pypi_0
terminado=0.9.3=py37h89c1867_0
testpath=0.4.4=py_0
threadpoolctl=2.1.0=pyh5ca1d4c_0
tifffile=2021.3.5=pypi_0
tk=8.6.10=hbc83047_0
tornado=6.1=py37h27cfd23_0
traitlets=5.0.5=pyhd3eb1b0_0
typing_extensions=3.7.4.3=py_0
tzdata=2020f=h52ac0ba_0
urllib3=1.26.3=pyhd8ed1ab_0
wcwidth=0.2.5=py_0
webcolors=1.11.1=pypi_0
webencodings=0.5.1=py_1
werkzeug=1.0.1=pypi_0
wheel=0.36.2=pyhd3eb1b0_0
widgetsnbextension=3.5.1=py37h89c1867_4
wrapt=1.12.1=pypi_0
x264=1!152.20180806=h14c3975_0
xorg-kbproto=1.0.7=h14c3975_1002
xorg-libice=1.0.10=h516909a_0
xorg-libsm=1.2.3=h84519dc_1000
xorg-libx11=1.6.12=h516909a_0
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xorg-libxext=1.3.4=h516909a_0
xorg-libxrender=0.9.10=h516909a_1002
xorg-renderproto=0.11.1=h14c3975_1002
xorg-xextproto=7.3.0=h14c3975_1002
xorg-xproto=7.0.31=h14c3975_1007
xz=5.2.5=h7b6447c_0
zeromq=4.3.3=he6710b0_3
zipp=3.4.1=pyhd8ed1ab_0
zlib=1.2.11=h7b6447c_3
zstd=1.4.5=h6597ccf_2
